requirements.txt
flask
flask-cors
torch
torchvision
faiss-cpu
pillow
numpy
pandas
reportlab
matplotlib
gunicorn
